@@ -65,50 +65,55 @@
     "objectpath": "^1.2.2",
     "prop-types": "^15.7.2",
     "rc-select": "^9.2.1",
-    "react": "^16.10.2",
+    "react": "^16.12.0",
     "react-ace": "^8.0.0",
-    "react-dom": "^16.10.2",
-    "react-dropzone": "^10.1.10",
+    "react-dom": "^16.12.0",
+    "react-dropzone": "^10.2.1",
     "reactcss": "^1.2.3",
-    "rxjs": "~6.5.2",
+    "rxjs": "~6.5.3",
     "schema-inspector": "^1.6.8",
-    "screenfull": "^4.2.1",
+    "screenfull": "^5.0.0",
     "split.js": "^1.5.11",
     "tinycolor2": "^1.4.1",
+    "tooltipster": "^4.2.7",
     "tslib": "^1.10.0",
     "tv4": "^1.3.0",
     "typeface-roboto": "^0.0.75",
     "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^8.2.0",
-    "@angular-devkit/build-angular": "^0.802.0",
-    "@angular/cli": "~8.2.2",
-    "@angular/compiler-cli": "~8.2.11",
-    "@angular/language-service": "~8.2.11",
+    "@angular-builders/custom-webpack": "^8.4.1",
+    "@angular-devkit/build-angular": "^0.803.20",
+    "@angular/cli": "~8.3.20",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/language-service": "~8.2.14",
     "@types/flot": "0.0.31",
-    "@types/jasmine": "~3.4.0",
-    "@types/jasminewd2": "~2.0.6",
+    "@types/jasmine": "~3.5.0",
+    "@types/jasminewd2": "~2.0.8",
     "@types/jquery": "^3.3.31",
     "@types/js-beautify": "^1.8.1",
-    "@types/node": "~10.14.15",
-    "@types/react": "^16.9.9",
-    "@types/react-dom": "^16.9.2",
+    "@types/node": "~12.12.17",
+    "@types/react": "^16.9.16",
+    "@types/react-dom": "^16.9.4",
     "@types/tinycolor2": "^1.4.2",
-    "codelyzer": "~5.1.0",
-    "compression-webpack-plugin": "^3.0.0",
-    "directory-tree": "^2.2.3",
-    "jasmine-core": "~3.4.0",
+    "@types/tooltipster": "0.0.29",
+    "codelyzer": "~5.2.0",
+    "compression-webpack-plugin": "^3.0.1",
+    "directory-tree": "^2.2.4",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.2.0",
-    "karma-chrome-launcher": "~3.0.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.1",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
     "ngrx-store-freeze": "^0.2.4",
     "protractor": "~5.4.2",
-    "ts-node": "~8.3.0",
-    "tslint": "~5.18.0",
+    "ts-node": "~8.5.4",
+    "tslint": "~5.20.1",
     "typescript": "~3.5.3"
+  },
+  "resolutions": {
+    "serialize-javascript": "^2.1.1"
   }
 }
