@@ -28,7 +28,7 @@ import { alarmFields } from '@shared/models/alarm.models';
 import { materialColors } from '@app/shared/models/material.models';
 import { WidgetInfo } from '@home/models/widget-component.models';
 import jsonSchemaDefaults from 'json-schema-defaults';
-import * as materialIconsCodepoints from '!raw-loader!material-design-icons/iconfont/codepoints';
+import materialIconsCodepoints from '!raw-loader!material-design-icons/iconfont/codepoints';
 import { Observable, of, ReplaySubject } from 'rxjs';
 
 const varsRegex = /\$\{([^}]*)\}/g;
